refactor(imports): migrate legacy normalize helpers to imports namespace
Move as 10 funções legadas de `apps/core/services/normalize.py` para o
namespace canônico `apps/core/imports/` (`normalization.py` + `hashing.py`).

`apps/core/services/normalize.py` vira **wrapper de compatibilidade**
(re-export apenas) — callers existentes (`controle_imports`,
`controle_acoes_import`, `dat_cadastros_import`, `views_lookup`) não
precisam mudar import lines.

## Funções movidas

- normalization.py: norm_text, normalize_sector, normalize_project_alias,
  split_municipios_super, parse_date_iso, parse_time_iso, normalize_email,
  normalize_date_field, normalize_time_field
- hashing.py: hash_event_v2

## Comportamento

Byte-equivalente. Os 75 snapshots de
`test_services_normalize_equivalence.py` (PR #1350) continuam verdes —
incluindo os quirks legados documentados (NaN literal, ISO-only date,
norm_text(None), etc).

Hash:
- algoritmo SHA1 preservado
- usedforsecurity=False preservado
- ordem dos 17 campos do hash_event_v2 preservada
- delimitador pipe preservado
- digests histórico válidos

## Não-mudanças

Endpoints, contratos, dry-run, RBAC, models, migrations, frontend e
sheets.banco não alterados.

Closes #1349
Depends on #1350
